refactor contract metadata view
